added mailing list field to users table
