+++ server/utils/proximity.js
